+++ examples/iam_v2/certificate_auth_providers_v2.yml
@@ -5,9 +5,9 @@
 # 2. Update certificate authentication provider
 # 3. List all certificate authentication providers
 # 4. Get specific certificate authentication provider
-# 5. Delete created certificate authentication providers
+# 5. Delete created certificate authentication provider
 
-- name: Certificate Authentication Providers playbook
+- name: Certificate Authentication Provider playbook
   hosts: localhost
   gather_facts: false
   module_defaults:
@@ -28,7 +28,7 @@
         directory_service_ext_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
 
     - name: Create certificate authentication provider with CAC enabled
-      nutanix.ncp.ntnx_certificate_auth_providers_v2:
+      nutanix.ncp.ntnx_certificate_auth_provider_v2:
         state: present
         name: "CAC"
         client_ca_chain: "{{ ca_chain_content }}"
@@ -44,7 +44,7 @@
         cert_auth_ext_id_cac: "{{ result_cac.ext_id }}"
 
     - name: Update certificate authentication provider
-      nutanix.ncp.ntnx_certificate_auth_providers_v2:
+      nutanix.ncp.ntnx_certificate_auth_provider_v2:
         state: present
         ext_id: "{{ cert_auth_ext_id_cac }}"
         is_cac_enabled: false
@@ -56,6 +56,13 @@
       register: result
       ignore_errors: true
 
+    - name: List all certificate authentication providers with page and limit
+      nutanix.ncp.ntnx_certificate_auth_providers_info_v2:
+        page: 0
+        limit: 1
+      register: result
+      ignore_errors: true
+
     - name: Get specific certificate authentication provider
       nutanix.ncp.ntnx_certificate_auth_providers_info_v2:
         ext_id: "{{ cert_auth_ext_id_cac }}"
@@ -63,7 +70,7 @@
       ignore_errors: true
 
     - name: Delete certificate authentication provider
-      nutanix.ncp.ntnx_certificate_auth_providers_v2:
+      nutanix.ncp.ntnx_certificate_auth_provider_v2:
         state: absent
         ext_id: "{{ cert_auth_ext_id_cac }}"
       register: result
